using libxxx.a instead of sepreated *.o files. (this exploration is hard,at least 5 hours were wasted on this matter)

+++ src/TSS.cpp
@@ -19,7 +19,7 @@ TSS::~TSS()
 
 void TSS::writeToMemory(int seg,int off)
 {   
-    Util::memcopy(Util::SEG_CURRENT,this->I0,seg,off,PMLoader::TSS_MIN_SIZE);
+    Util::memcopy(Util::SEG_CURRENT,(int)this->I0,seg,off,PMLoader::TSS_MIN_SIZE);
 }
 void TSS::ensureReservedZero()
 {
@@ -36,7 +36,7 @@ void TSS::ensureReservedZero()
 }
 void TSS::fromMemory(TSS &self,int seg,int off)
 {
-    Util::memcopy(seg,off,Util::SEG_CURRENT,self.I0,PMLoader::TSS_MIN_SIZE);
+    Util::memcopy(seg,off,Util::SEG_CURRENT,(int)self.I0,PMLoader::TSS_MIN_SIZE);
     self.ensureReservedZero();//保证所有的reserved都为0
 }
